[PHP] Address indentation for embedded CSS and JS
To do this, the JS and CSS contexts must not be pushed on the same line
as the opening tag, if it only consists of whitespace, so that the
scope at the end of the line is still `text.html.php`.

In order to allow for a more consistent application of the embedded
scope names, we also embed JS and CSS on the first newline we
encounter.