[Git Formats] Git Attributes maintanance (#2802)
* [Git Formats] Fix context name in Git Attributes

This commit renames `else-pop` context to `immediately-pop` to comply
with scope names used in other syntaxes.

* [Git Formats] Exclude prototypes from meta-contexts in Git Attributes

This commit explicitly excludes prototype context from contexts, which
exist to just add a meta scope and immediately pop. They are not meant
to receive other patterns in case of being inherited.

* [Git Formats] Add named string-content context in Git Attributes

This commit adds a named context `string-content` for patterns of quoted
strings. It also adds `meta.string` scopes to comply with other syntaxes.