[Markdown] Fix refs/links after footnotes
This commit fixes a syntax highlighting issue which causes inline links
or references not to be highlighted correctly if those directly follow
a footnote.

The footnote was highlighted as reference causing the reference
description to be highlighted as url. The reference's url was not
highlighted at all.

The reason is footnotes not having an url-part. So they must be matched
before literal references.