[Markdown] Remove keybinding for fenced codeblocks
It appears the removed key binding behaves different depending on
keyboard layout (deadkeys). This results in sub-optimal behavior for a
reasonable number of users.

It may also jump in if one tries to add inline codeblocks at the
beginning of a line.

Attempts to bind `["`", "`", "`"] or tweak `context` have not been
successful.

Adding syntax specific settings to gate this binding are probably not
worth it.

Hence it's probably better to remove it.

see:
- https://forum.sublimetext.com/t/markdown-fenced-code-snippet/58261/3
- https://discord.com/channels/280102180189634562/280157083356233728/866850636083494963