[HTML] Simplify CSS/JS embedding
This commit removes one level of abstraction.

The removed separation between  `...-begin` and `...-content` context
turned out to be useless with regards to avoid dependencies between
inherited and inheriting syntax definition.

It's because the inheriting syntax needs to handle those HTML comments
in front of closing tags by itself anyway. So it can do the same for
the opening comment and we can save one useless context switch.

Learned from adding VBScript highlighting in script tags of ASP syntax.