[CSS] Indent lines with leading block comments
This commit makes sure to handle lines with leading block comments as
normal code, which means they are handled by indentation engine.

Before this commit, those lines were ignored.

What's a bit unclear at the moment is, why those lines are aligned to
the beginning of the line. Maybe because the engine ignores the comment
scope?