[Markdown] Terminate inline formatting on paragraph boundaries (#3036)
This commit ...

1. includes `paragraph-end` into `bold-italic-trailing` context
2. moves the `bold-italic-trailing` includes before `inline` includes
   to terminate inline formatting by fenced code blocks
3. renames `bold-italic-trailing` to `bold-italic-common` and moves the
   `inline` include into it.

Everything, which terminates a paragraph, now also terminates inline
formatting.

Note:

A token, which starts with an inline-formatting character such as `_` or
`*` but is terminated by a paragraph boundary is not rendered bold/
italic/strikethrough by CommonMark.

It is still scoped/highlighted as such by this syntax. This commit only
fixes paragraph boundaries which have been ignored in such situations
before.

A future commit may work on fixing this by using branch points, but it
requires way more considerations and a possibly larger overhaul of
inline formatting contexts.