[Markdown] Faster code fence escape
This commit makes `code_fence_end` consume the newline character, which
results in slightly faster parsing as it doesn't leave anything to
match when returning to `fenced-code-block-content`. As a result it is
popped faster.

Note: This change is not directly related with the content of this PR
      but the difference in behavior was found when comparing
      `code_fence_end` and the pattern from `fenced-raw` context, which
      already consumes the newline optionally.

It doesn't change behavior except those tiny impact on performance.