[Markdown] Highlight any unknown fenced language identifier (#3037)
* [Markdown] Highlight any unknown fenced language identifier

This commit implements a re-interpreted version of CommonMark's
statement:

  The first word of the info string is typically used to specify
  the language of the code sample...

  see: https://spec.commonmark.org/0.30/#fenced-code-blocks

A word is not only `[\w-]+` but a sequence of any non-whitespace
character, except backticks, in this context.

This is why `C++` is an valid fenced code language identifier.

This commit expects a language identifier to start with a letter,
in order to avoid confusion with e.g. attribute annotations such as
`{key=value}` which some Markdown extension support.

* [Markdown] Faster code fence escape

This commit makes `code_fence_end` consume the newline character, which
results in slightly faster parsing as it doesn't leave anything to
match when returning to `fenced-code-block-content`. As a result it is
popped faster.

Note: This change is not directly related with the content of this PR
      but the difference in behavior was found when comparing
      `code_fence_end` and the pattern from `fenced-raw` context, which
      already consumes the newline optionally.

It doesn't change behavior except those tiny impact on performance.