[Markdown] Restrict image/link/reference highlighting
See: https://github.com/SublimeText-Markdown/MarkdownEditing/issues/633

Space between description and metadata is not allowed in inline links,
images or references. Markdown must therefore not claim it to be such an
entity as it causes otherwise valid markdown files to get highlighted
wrong.

Example:

   [description] (no more url)

This is no inline link according to CommonMark or GFM.