[Python] Fix lambdas inside list comprehensions
Fixes #3049

This commit adds a dedicated `lambda-in-groups` context, which differs
from normal `lambda` only by an additional bailout to pop contexts off
stack in case of `for` keyword being matched.

Let the including context handle/consume it.

Duplication of contexts is required to maintain illegal highlighting of
`for` keyword in top-level statements, if not surrounded by brackets.