[TypeScript] Support export declare enum/export abstract class. (#3061)
Fix #3060.

The `abstract` and `declare` keywords were only handled in `statement`, not in the export stuff. This adds them to `export-extended`.

It might be a good idea to move all of the exportable declarations into a single new context in the future.

Also, the scopes on `enum` and `declare` should probably be updated in the future.