[TypeScript] Fix a bug in function type parsing. (#3062)
Fix https://github.com/sublimehq/Packages/issues/3059#issue-1011995899.

Function types are parsed similarly to arrow functions — the content of the
parentheses is parsed as a parenthesized type, and then if an arrow is found
after the close paren, the parser backtracks and re-parses as a function type.
But if parsing goes far wrong within the parens, then it can't find the arrow.
This can happen because function type parameter lists are not necessarily valid
parenthesized types. For arrow function literals, we account for this by making
sure that the expression grammar “covers” the parameter list grammar. That
doesn't work here because we need to be very strict with bailing out of invalid
types. Instead, when we see that a parenthesized type has gone wrong, we
immediately backtrack and try to reparse as a function type.

This shouldn't affect any previously-valid syntax, but it might result in
different behavior for invalid syntax. Probably it should result in better
behavior, since it's more likely that a function type will be identified early
while typing the argument list.