[Python] Fix highlighting of match with type annotation
This commit fixes an issue which caused a variable named `match` being
scoped as keyword if it is followed by an annotation.

This is achieved by failing the match-statement branch if the `match`
token is directly followed by a colon.

Note: The `match` token may be scoped keyword only if a valid match
      statement is identified. Absence of a variable to match
      invalidates it.