[Python] Add shebang
This commit adds a dedicated context to scope the shebang line.

Some fonts like provide ligatures for the `#!` sign, thus need to
ensure it is tokenized correctly.

The statements context is set onto stack if shebang matches but not
after the next start of line or non-whitespace character. This is to
make sure, the shebang is highlighted within fenced code blocks of
Markdown.