[Matlab] Add code section comments to symbol list (#3219)
Reference from the Matlab docs: https://www.mathworks.com/help/matlab/matlab_prog/create-and-run-sections.html

The scope `entity.name.section` is added to the section title, which is already used in R and gets a "navigation" kind icon in the Goto Symbol panel. Also the pattern was adjusted so that section comments should start on their own line.