[PHP] Suppress method completions when typing php tags (#3237)
* [PHP] Suppress method completions when typing php tags

Resolves https://github.com/sublimehq/sublime_text/issues/5256

The fix consists of two parts:

1. Extending the completion's selector by 

       - punctuation.section.embedded.begin

2. Making sure to highlight incomplete <?php tags

       punctuation.section.embedded.begin
    
   Note: If this isn't acceptable we might want to scope incomplete
   tags using a `meta` scope and add this to the selector (1.) as well.

   We just need to ensure ST sees the correct scope while typing the
   opening php tag in order to not add those methods to the list of
   available completions.

* [PHP] Fix php snippet

This commit makes sure to correctly complete the `<?php | ?>` snippet
depending on whether `php` or `<?php` was entered. In the latter case
we might want to apply `php (begin tag)` snippet which doesn't add a
second `<?`.

Before this commit:

  <?ph|   =>  <?<?php | ?>

With this commit:

  <?ph|   =>  <?php | ?>

* [PHP] Use separate patterns for incomplete php tags

Keep scoping to suppress completions while not pushing/popping contexts
until tag is complete. May help to prevent some false positives or
confusion about non-running but highlighted code.