[Python] Fix implicit tuples termination
Fixes #3243

Pop `allow-unpack-operators` off stack directly at the beginning of a
line if it doesn't begin with optional whitespace followed by asterisk
or hashtag.

This is to ensure to correctly match `def` statements whose pattern
requires to be matched at bol.

The new pattern is appended to keep performance impact as low as
possible. It is matched upon newlines only.