[PHP] Rename function-call-parameters
Instantiations use the term `arguments`. Several other syntaxes have
moved on to use the term `parameters` in declarations and `arguments`
when invoking a function. The title of a referenced specification
(https://wiki.php.net/rfc/named_params) uses `named arguments`.

So even not directly related with fixing anonymous class instantiation,
this commit renames related contexts to match the found naming scheme.
It's put in here as function-call contexts needed some tweaks to be
reused in instantiations anyway.

Note: This commit doesn't change behavior of the syntax.