[Markdown] Add support for setext headings in list-blocks (#3399)
* [Markdown] Add support for setext headings in list-blocks

This commit adds `list-setext-headings-or-paragraphs` contexts which
are a modified copy of `setext-headings-or-paragraphs` to support
setext headings within list items.

It may also be useful if markdown is to be embedded into 3rd-party
syntaxes with lazy indentation handling.

* [Markdown] Tweak setext heading escapes

1. `setext-heading1` is the last branch, doesn't need a failure.
2. `setext-heading2` needs to fail on setext heading level 1 only.

* [Markdown] sync setext trailing whitespace patterns

Both ..._escape and ..._end patterns should match the same style of
trailing whitespace to avoid theoretical bugs with invisible unicode
spaces.

Second change causes leading whitespace at the beginning of a pragraph
not to be consumed by `setext_heading_or_paragraph` pattern. This will
be required for mathjax blocks.

* [Markdown] Reorg paragraph-end vs. -fail

Goal is same structure of both heading/paragraph sections and to make
`fail` pattern available for use by other contexts.

* [Markdown] Tweak setext patterns in lists

This commit adds a heuristic to require at least 2 hyphens or equal
signs for setext headings in list blocks in order to prevent items
followed by empty items being scoped heading temporarily.

The heuristic is required as we can't track indentation.