[Java] Fix default method modifier scope
Fixes #3438

This commit re-orders `declarations` and `statements` contexts to ensure to first consume `default Type method()` declarations, which is
required to properly scope `default` keyword in that context.

That change revealed `yield` being missing in `control_keywords` which
caused it to be scoped `storage.type` in some contexts by accident.
That's fixed to keep syntax tests passing.