[Java] Add dedicated context for instanceof expressions
This commit adds `type-comparisons` context to implement the
`instanceof { TypePattern }` expression in the TYPE PATTERNS section as
it belongs there. It's the main use case for type patterns. As a result
`operators` context contains symbolic operators only as of this commit.

Note: This commit doesn't change behavior of Java syntax.