[PHP] Tweak type-hint performance
This commit adjusts type hint implementation to avoid some redundant
lookaheads once a type hint context is on stack.

The new test case is meant to ensure not to break reference operator
precedence in function parameter lists as this is the only change in
behavior this commit includes.

Benchmark:

10k lines of:

   function unionTypeFunction(int | string $param0, Foo | \Foo\Bar | ?int $param1, Foo|\Foo\Bar|?int $param2, (?A|B)&(C|?D)&?E $param2, string $param3, $param4): Foo|\Foo\Bar|?int|static {}

Result:

   -5% parsing time