[Rust] Change scopes of `!` and `?` type (modifiers)
`!` as the "Never" type is a proper type and should be matched as such.
https://doc.rust-lang.org/reference/types/never.html

`?` and `!` as prefixes to Sized, Send, or Sync modify the given type
name and provide more information to them. Thus they should receive a
`storage.modified` scope instead of `keyword.operator`.

Also make these modifiers available for all types and not just a
hard-coded list because the compiler will complain if it's not valid
there.

Note that `impl !AutoTrait for T {}` is exempted from this because here
the `!` is relevant to meaning of the impl block in that it prevents
the creation of auto-trait implementations.
https://doc.rust-lang.org/reference/special-types-and-traits.html#auto-traits