[Rust] Specify `?` and `!` punctuation scopes
`?` is a special marker for relaxing the points of `Sized` and currently
only allowed to be used with `Sized` [1], but it's not necessary for us to
make this restriction (and actually complicates things).

`!` as a modifier is only used to opt-out of trait auto-implementations,
which makes it interesting only in places where a trait is specified.
There was already a proper match for this in `impl` statements and used
a `keyword.operator` there, but for `trait` definitions that was not
the case and needed the fallback in `type-any-identifiers`.
It's also called "negative impl" in the rustonomicon. [2]

Also fix `!` being used before a path, as done in `Rc` of the std lib.

[1]: https://www.uditsamani.com/sized-in-rust/#pro-tips-1
[2]: https://doc.rust-lang.org/nomicon/send-and-sync.html