[Rust] Various tweaks and fixes to type matching (#3510)
* [Rust] Remove double and incorrect `storage` scope

The accessor punctuation used to also be part of this scope when it
should not have been.

* [Rust] Unify indentation

* [Rust] Restructure type matching

- Introduce a separate more relaxed context for matching types with
  higher confidence.
- Fix paths not being recognizes as such when the first identifier is a
  known name such as `Self` or `Iterator`.
- Fix non-known types not being highlighted
  * within a known generic type,
  * within hinted function calls,
  * within function signatures.
- Remove double include of `support-type` in `statement` context.
- Rename some type contexts to follow naming conventions.

* [Rust] Change scopes of `!` and `?` type (modifiers)

`!` as the "Never" type is a proper type and should be matched as such.
https://doc.rust-lang.org/reference/types/never.html

`?` and `!` as prefixes to Sized, Send, or Sync modify the given type
name and provide more information to them. Thus they should receive a
`storage.modified` scope instead of `keyword.operator`.

Also make these modifiers available for all types and not just a
hard-coded list because the compiler will complain if it's not valid
there.

Note that `impl !AutoTrait for T {}` is exempted from this because here
the `!` is relevant to meaning of the impl block in that it prevents
the creation of auto-trait implementations.
https://doc.rust-lang.org/reference/special-types-and-traits.html#auto-traits

* [Rust] Specify `?` and `!` punctuation scopes

`?` is a special marker for relaxing the points of `Sized` and currently
only allowed to be used with `Sized` [1], but it's not necessary for us to
make this restriction (and actually complicates things).

`!` as a modifier is only used to opt-out of trait auto-implementations,
which makes it interesting only in places where a trait is specified.
There was already a proper match for this in `impl` statements and used
a `keyword.operator` there, but for `trait` definitions that was not
the case and needed the fallback in `type-any-identifiers`.
It's also called "negative impl" in the rustonomicon. [2]

Also fix `!` being used before a path, as done in `Rc` of the std lib.

[1]: https://www.uditsamani.com/sized-in-rust/#pro-tips-1
[2]: https://doc.rust-lang.org/nomicon/send-and-sync.html

* [Rust] Rename scope for `_` type placeholder