[TeX] extended control flow patterns and scopes (#3533)
This commit adds fine-grained scoping for control structures in TeX.

Q: How important is it that these scopes are balanced? In the examples in the test, 
the `keyword.control.conditional.if.tex` never has any corresponding `end` scope, because the loop-ending `repeat` closes both the `loop`, and the `if` as part of the loop condition.

And the conditions `\if@director` would be using a custom-defined if condition. So we have `else` and `fi` without a corresponding start of that scope. I'm not sure if this can be fixed. Maybe we can assume that all commands that start with the sequence `\if` and are not followed by brace-enclosed arguments probably constitute the beginning of an `if`, but I'm not sure how many false positives (or false negatives) this would bring.