[JavaScript] Use branching to improve continuation scopes. (#3257)
When a decorator ends at the end of a line, end the annotation scope there. 
This requires branching to look across the line boundary.

Notes:

Also reimplement expression statement line continuation in the same way. 
This eliminates the old set/set back pattern in favor of branch/pop. 
I was going to reuse this pattern in `variable-binding-list-top`, 
but it turns out that the line continuation code there wasn't doing 
anything in the first place, so I removed it.