[TeX] added scoping of "constants" (#3581)
This commit adds scoping/highlighting for the built-in constant-like objects in TeX.

The decision which control sequences to include here is based more on expected
usage, than on how this is defined in the underlying language. I think for a 
macro-processing system like TeX, this is probably the best approach. 

Thus I've considered as constants in this sense:
 - The underlying parameters of the TeX typesetting algorithm. 
 - Registers with pre-defined meaning
 - simple macros (i.e. without parameters) that insert a fixed element into
   the typesetting
 
In particular, a parameterless macro that 
a) assigns to some internal value, or 
b) changes how subsequent input tokens will be interpreted are not considered 
   constants here.

The typesetting parameters and registers can be assigned-to, though for most of
these this would typically only happen once in the beginning of the document
(usually hidden from the user, as part of the TeX format that is used).
However, it is possible to change these values in the middle of the document
(though some **really** should not be changed), so in that sense these are not
all technically constants. 

For the assignment, I've opted to not try to distinguish constants that are
dimensions from constants that are pure numbers, so both get the same scoping.
In extremely weird corner cases, this might lead to erroneous scoping, e.g. if
you write
```
\hyphenpenalty=5pt
```
then the current scoping would not know that the penalty
is a number, so the `pt` are actually letters to typeset, and not part of the
unit. Such cases should be very rare, though.

Similarly, since macros can hide part of the syntax, there are situations 
where we have no chance of knowing which scoping is correct. For example, in
```
\abovedisplayskip\value plus 5pt
```
if `\value` evaluates to some measurement, then the `plus 5pt` would be part 
of the measurement, otherwise, it is normal text.

Co-authored-by: deathaxe <deathaxe82@googlemail.com>