[TeX] Improved integer handling (#3604)
This commit does a first step at generalizing and unifying the handling of integers.
It provides a general context that matches decimal, octal, hexadecimal, and 
"based on character" integers. I've not yet used this for the definition of dimensions, 
because here we currently enter different scopes based on whether we've matched
an integer or a float. The syntax specified here assumes that everything will be
written on one line. It also does not handle the case of multiple leading signs.