Drop custom build target
This commit uses default ExecCommand to run a shell script.

By running it via `shell_cmd` a login shell is invoked which evaluates
the shebang. Hence it's no longer needed to be done by an extra plugin.

On Windows MSYS2's `sh` command is used to invoke a shell command.
It requires it to be present on %PATH% to work, but before this PR
Windows was not supported at all.