[LaTeX] Text formatting (#3618)
This commit reworks the handling of text formatting commands in LaTeX. 
In particular, in cases where there are commands within formatted text, 
e.g. `\textbf{\TeX in bold}`, we no longer format the command names. 

One challenge is that there are several font properties that can be changed
independently:

- Font family (roman, sans serif, typewriter), 
- font weight (bold, medium, (possibly) light), 
- and font shape (upright, italic, slanted, small-caps). 

Most of these do not have any corresponding `markup` scopes and highlighting
rules (at least with default color schemes). At this point, only a subset of
nesting of these properties is actually handled: Bold/italic combination, 
and nested \emph emphasize, as well as ensuring that all the commands that
do not change one of these properties still typeset their arguments according
to the outer command.

No provision has been made for any form of triple-nesting, 
i.e. `\textbf{\textit{\textup{a}}}` would not result in bold `a` right now. 
Not sure if supporting this would be worth it.

Finally, there is also an \underline command that could be combined with bold/italic. 
So far, I've not included this, because in LaTeX underline comes with some 
caveats, so that package commands like `\uline` or `\ul` are often recommended.