[JavaScript] Fix syntax engine deadlock (#3656)
* [JavaScript] Fix syntax engine deadlock

Works around: 
- https://github.com/sublimehq/sublime_text/issues/5813
- https://github.com/sublimehq/sublime_text/issues/5853

ST locks up on certain text manipulation operations, which seem to cause
syntax engine to bring into an invalid state causing a deadlock or
infinite loop.

The two related branch points are used to peak into the next line to
decide whether the current context's expression continues. If not the
current context is to be popped off stack (immediately).

"Immediately" is what causes ST to lock up, on certain text
manipulations if the context/branch is nested in another branch.

This commit works around this bug by popping the fallback branch at the
same text point as the main branch failed at, using the same pattern
to pop. As a result tokens which are matched by `{{line_ending_ahead}}`
need to be consumed by the fallback context. In that case these are
comments. A possible meta scope needs to be cleared and prototypes from
possible inheriting syntaxes are to be skipped.

* [JavaScript] Exclude trailing line comments from annotations

* [JavaScript] Reformat expression continuation pattern

* [JavaScript] Add else-pop-2 prototype context