[Matlab] Add heuristic to pop function context (#3658)
This commit pops the `meta.function` scope when another `function`
keyword appears at the start of a line, which is unlikely to indicate a
nested function. It is meant to prevent the accumulation of
`meta.function` scopes if a files contains many local functions without
the `end` keyword, which is not required under certain conditions.