[CSS] Normalize property-name contexts
This commit normalizes property name behavior by directly checking for
property values after a name. This has already be the case for `font`,
`font-family` and other special treated properties. It is now done for
all of them.

This change is not directly needed by CSS, but helps 3rd-party syntaxes
such as PostCSS, TailwindCSS, LESS, SCSS to implement nested rule sets.

This commit does not have any effect on how CSS is highlighted.