[CSS] Introduce value-prototype context
This commit adds an organizational context, which most/all value related
expressions have in common. It can also be used as a common entry point
for extensions added by inheriting syntaxes.

The `builtin-values` and `common-operators` contexts are merged into the
new `values` context as they seem to provide little to no use anymore.