[Batch File] Improve batch file matching
This commit aims to improve batch file (i.e. `cmd.exe` script) matching by:

  - adding `.bat` and `.cmd` to the list of valid file extensions
    (dotfile scripts e.g. `%USERPROFILE%\.{bat,cmd}` invoked via
    `{HKLM,HKCU}\Software\Microsoft\Command Processor\AutoRun`);
  - enhancing `first_line_match` in order for it to match most of
    the common first line patterns in batch files, specifically
    "ifdebugecho" (`@if "%DEBUG%"=="" @echo off`) and its variants.

An effort was made to make it a catch-all; however, it turned out to be on par
with Herculean trials - there are unfathomable, unspeakable things lurking
in these scripts - trying to predict them all is an impossible task.

(Ever heard of a batch file embedding a Python script?)