[Go] Avoid including `main`
As soon as a syntax wants to support shebang or frontmatter, including
main within normal statements is not possible any longer.

Hence it is probably a general good strategy to not do that at all.
As Go's `main` context includes `match-any` only, it is the context
which seems more appropriate to use.