[Erlang] Remove obsolete query keyword (#3733)
Fixes #2401

This PR proposes to remove the obsolete keyword `query` from Erlang syntax, 
as it has been deprecated in 2007 and removed from Erlang-OTP in 2012. The atom
`query` can be used as normal variable/constant in any context, these days.

The last attempt to do so (see: #2409) was rejected because of backward 
compatibility concerns. With inheritance it is however trivial to create
a syntax, which re-adds the keyword if needed.

It would look like:

```yaml
%YAML 1.2
---
# https://www.erlang.org/doc/reference_manual
# https://www.sublimetext.com/docs/syntax.html
name: Erlang (Legacy)
scope: source.erlang
version: 2

extends: Erlang.sublime-syntax

contexts:
  expr-control:
    - meta_prepend: true
    # https://www.erlang.org/doc/reference_manual/...?
    - match: query{{ident_break}}
      scope: keyword.control.flow.query.erlang
      push:
        - meta_scope: meta.query.erlang
        - include: expr-control-flow-end
        - include: expr-control-body
```