[Python] Support normal expressions in annotations
This commit partly reverts restrictions to allow only type-hints in annotations.

For sake of consistent highlighting brackets are scoped `meta.brackets` instead
of `meta.item-access` and `meta.sequence.list.members`, so chances are lower for
equally looking terms to be higlighted differently.

Annotations are still scoped `meta.type` as this is the primary use case and the
scope is to be used by color schemes for dimming type hints easily.