[Batch File] Fix escaped percentage not being detected
This commit fixes an issue which caused a path pattern to be popped off stack
too early, because of an heuristic used in `for /r [path] %%f` statements to
make sure to match `%%f` as for variable if `[path]` is absent.

The heuristic is now restricted to for statements only, instead of being applied
to all path patterns globally.