[CSS] Highlight calc() in all function arguments
Fixes #3748

This commit adds `calc()` function to `function-arguments-prototype` so it is
highlighted in all function arguments. The function is valid in majority of
function calls. Maybe not everywhere, but we don't want to be a linter - or at
least reduce those habits.

A first step to a somewhat more lazy parsing of CSS.