[Python] Terminate parameter annotation before newline
This commit works around ST's line blindness via branching in order to correctly
terminate `meta.type` before `\n`, if annotation is terminated at one of the
following lines.

As for decorators, this is to make sure an applied background color doesn't tint
the whole line up to a view's right most boundary.