[Regular Expressions] Re-introduce file/path wildcards
ST treats `*` depending on whether being used in a path with `/` or in simple
file patterns. This commit re-introduces distinct scopes based on that, to
enable them being highlighted differently.

Note:

The old and resurrected behavior does not yet apply expected scope to patterns
like `fo*der/`. A fix is however out of scope for this PR.