[Python] Add syntax support for sphinx-style doc comments
Sphinx (the most common Python documentation tool) uses `#:` comments to
signal documentation comments for its "autodoc" module.

There is no known overlap with other comment punctuation, so this should
be a safe addition.

`TM_COMMENT_START_2` is also added to be able to undo the comment using
`toggle_comment`, but unfortunately it does not work like this
currently because `Default/comment.py` does not prioritize the longest
prefix match of all available comment styles but we don't want this
style to be the default.

https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#directive-autoattribute