[Rust] Restructure definition file
Move some context to related ones and add section headers.

This commit does not claim to be an accurate grouping, but some grouping
for now is better than none.

Additionally, the `statements-block` context was replaced with a `block`
context that pops after matching, following the plural vs singular
naming convention.