[TypeScript] Keyword types can't be generic (#3863)
Fix #3862.

Originally, type arguments were part of the regular type expression
structure, so the syntax would try to highlight type arguments after
any type. But actually, only ordinary named types can get them. In the
linked issue, this was causing problems with `as`-style type
assertions — when followed by a `<`, the type expression was eating it
and trying to parse type arguments, even when the preceding type does
not accept them.

According to the spec — I'm kidding, we all know the drill by now,
actually after a bunch of futzing around in AST Explorer — this should
work correctly in all cases I can think of.

In order for this to work in all cases, it's necessary to ensure that
all specially keyworded types are accounted for, so this PR also adds
the `true` and `false` singleton types.