[C#] Drop `NAnt Build File` syntax definition
It is basically a very rudimentary XML syntax that does not bring
anything to the table, does not have tests and does not have any other
integration within ST either. Because of the generic file extension,
and 100% overlap with the `first_line_match`, it cannot even be
reasonably used as a syntax definition by default and compared to the
standard XML syntax definition, which would automatically be selected
in 99% of all ST environments, it's just worse in every way.

See also #3569 for a previous discussion.