[Rust] Support macro repetitions inside structs
This only addresses a specific case of macro repetitions inside the
struct body and similar changes to other statements with body blocks
are likely also necessary, but I'm keeping the change to a limited in
order to observe its effects first. Further reports of such issues can
be handled analogously.