[Python] Improve type alias definitions
Fixes #3902

This commit ensures to highlight `type` more context sensitively.

It may have the following meanings:

1. built-in `type()` function
2. type alias declaration keyword: `type Alias`
3. ordinary variable used in many data classes

The built-in function (1.) most likely appears in r-value expressions, such as

   `my_type = type(var)`

At beginning of statements this commit prefers scoping `type` as keyword (2.)
or variable (3.) as this is the most likely use case.

The implementation ensures to start `meta` scope of alias declarations at bol
as class or function declarations do.